requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.9.0
click==8.1.3
colorama==0.4.6
decouple==0.0.7
Deprecated==1.2.14
dnspython==2.3.0
Flask==2.3.2
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Migrate==4.0.4
Flask-PyMongo==2.3.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mongoengine==0.27.0
ordered-set==4.1.0
packaging==23.1
pluggy==1.2.0
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.4.0
pyrsistent==0.19.3
pytest==7.4.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
rich==13.4.2
SQLAlchemy==2.0.17
typing_extensions==4.6.3
Werkzeug==2.3.6
wrapt==1.15.0
